pyproject.toml
[project]
name = "todo-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arize-phoenix-otel>=0.12.1",
    "openai-agents>=0.1.0",
    "openai>=1.93.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "weave>=0.51.54",
    "arize-phoenix>=11.2.0",
    "openinference-instrumentation-openai-agents>=0.1.14",
]
